requirements.txt
fastapi
uvicorn
neo4j
pandas
scikit-learn
torch
joblib
pydantic-settings
google-generativeai
